+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ponder/react
 
+## 0.12.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.8
+
+## 0.12.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.7
+
 ## 0.12.6
 
 ### Patch Changes